# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PlayerInputColor = typeof PlayerInputColor[keyof typeof PlayerInputColor];


export const PlayerInputColor = {
  red: 'red',
  green: 'green',
  yellow: 'yellow',
  blue: 'blue',
} as const;

export type PlayerInputLanguage = typeof PlayerInputLanguage[keyof typeof PlayerInputLanguage];


export const PlayerInputLanguage = {
  de: 'de',
  en: 'en',
  sq: 'sq',
} as const;

export interface PlayerInput {
  /**
     * @minLength 1
     * @maxLength 24
     */
  name: string;
  color: PlayerInputColor;
  language: PlayerInputLanguage;
  /**
     * @maxLength 1000000
     * @nullable
     */
  image?: string | null;
}

export interface CreateRoomInput {
  /**
     * @minimum 2
     * @maximum 4
     */
  targetCount: number;
  player: PlayerInput;
}

export interface JoinRoomInput {
  player: PlayerInput;
}

export type PlayerUpdateInputColor = typeof PlayerUpdateInputColor[keyof typeof PlayerUpdateInputColor];


export const PlayerUpdateInputColor = {
  red: 'red',
  green: 'green',
  yellow: 'yellow',
  blue: 'blue',
} as const;

export type PlayerUpdateInputLanguage = typeof PlayerUpdateInputLanguage[keyof typeof PlayerUpdateInputLanguage];


export const PlayerUpdateInputLanguage = {
  de: 'de',
  en: 'en',
  sq: 'sq',
} as const;

export interface PlayerUpdateInput {
  /**
     * @minLength 1
     * @maxLength 24
     */
  name: string;
  color: PlayerUpdateInputColor;
  language: PlayerUpdateInputLanguage;
  /**
     * @maxLength 1000000
     * @nullable
     */
  image?: string | null;
}

export interface ReadyInput {
  ready: boolean;
}

export interface MoveInput {
  /**
     * @minimum 0
     * @maximum 3
     */
  tokenIndex: number;
}

export type PlayerState = PlayerInput & {
  id: string;
  host: boolean;
  ready: boolean;
  tokens: number[];
  finished: number;
  connected: boolean;
};

export type RoomStateGamePhase = typeof RoomStateGamePhase[keyof typeof RoomStateGamePhase];


export const RoomStateGamePhase = {
  lobby: 'lobby',
  rolling: 'rolling',
  moving: 'moving',
  finished: 'finished',
} as const;

export interface RoomState {
  roomId: string;
  hostId: string;
  targetCount: number;
  players: PlayerState[];
  /** @nullable */
  currentPlayer: string | null;
  /** @nullable */
  diceValue: number | null;
  gamePhase: RoomStateGamePhase;
  /** @nullable */
  winner: string | null;
  gameStarted: boolean;
  version: number;
}

export interface JoinRoomResponse {
  playerId: string;
  state: RoomState;
}

export type RoomCodeQueryParameter = string;

export type PlayerIdQueryParameter = string;

export type JoinLudoRoomParams = {
/**
 * @pattern ^[0-9]{6}$
 */
code: RoomCodeQueryParameter;
};

export type GetLudoRoomParams = {
/**
 * @pattern ^[0-9]{6}$
 */
code: RoomCodeQueryParameter;
playerId: PlayerIdQueryParameter;
};

export type UpdateLudoPlayerParams = {
/**
 * @pattern ^[0-9]{6}$
 */
code: RoomCodeQueryParameter;
playerId: PlayerIdQueryParameter;
};

export type SetLudoReadyParams = {
/**
 * @pattern ^[0-9]{6}$
 */
code: RoomCodeQueryParameter;
playerId: PlayerIdQueryParameter;
};

export type RollLudoDiceParams = {
/**
 * @pattern ^[0-9]{6}$
 */
code: RoomCodeQueryParameter;
playerId: PlayerIdQueryParameter;
};

export type MoveLudoPieceParams = {
/**
 * @pattern ^[0-9]{6}$
 */
code: RoomCodeQueryParameter;
playerId: PlayerIdQueryParameter;
};