# lib/api-zod/src/generated/types/createRoomInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerInput } from './playerInput';

export interface CreateRoomInput {
  /**
     * @minimum 2
     * @maximum 4
     */
  targetCount: number;
  player: PlayerInput;
}