# lib/api-zod/src/generated/types/joinRoomResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomState } from './roomState';

export interface JoinRoomResponse {
  playerId: string;
  state: RoomState;
}