# lib/api-zod/src/generated/types/moveInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MoveInput {
  /**
     * @minimum 0
     * @maximum 3
     */
  tokenIndex: number;
}