# lib/api-zod/src/generated/types/playerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerInputColor } from './playerInputColor';
import type { PlayerInputLanguage } from './playerInputLanguage';

export interface PlayerInput {
  /**
     * @minLength 1
     * @maxLength 24
     */
  name: string;
  color: PlayerInputColor;
  language: PlayerInputLanguage;
  /**
     * @maxLength 1000000
     * @nullable
     */
  image?: string | null;
}