# lib/api-zod/src/generated/types/playerState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerInput } from './playerInput';

export type PlayerState = PlayerInput & {
  id: string;
  host: boolean;
  ready: boolean;
  tokens: number[];
  finished: number;
  connected: boolean;
};