# lib/api-zod/src/generated/types/playerUpdateInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerUpdateInputColor } from './playerUpdateInputColor';
import type { PlayerUpdateInputLanguage } from './playerUpdateInputLanguage';

export interface PlayerUpdateInput {
  /**
     * @minLength 1
     * @maxLength 24
     */
  name: string;
  color: PlayerUpdateInputColor;
  language: PlayerUpdateInputLanguage;
  /**
     * @maxLength 1000000
     * @nullable
     */
  image?: string | null;
}