# lib/api-zod/src/generated/types/roomState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerState } from './playerState';
import type { RoomStateGamePhase } from './roomStateGamePhase';

export interface RoomState {
  roomId: string;
  hostId: string;
  targetCount: number;
  players: PlayerState[];
  /** @nullable */
  currentPlayer: string | null;
  /** @nullable */
  diceValue: number | null;
  gamePhase: RoomStateGamePhase;
  /** @nullable */
  winner: string | null;
  gameStarted: boolean;
  version: number;
}